modify packetType

+++ src/main/java/com/dreamgyf/gmqyttf/common/enums/MqttPacketType.java
@@ -1,18 +1,18 @@
 package com.dreamgyf.gmqyttf.common.enums;
 
 public interface MqttPacketType {
-    int CONNECT = 1;
-    int CONNACK = 2;
-    int PUBLISH = 3;
-    int PUBACK = 4;
-    int PUBREC = 5;
-    int PUBREL = 6;
-    int PUBCOMP = 7;
-    int SUBSCRIBE = 8;
-    int SUBACK = 9;
-    int UNSUBSCRIBE = 0;
-    int UNSUBACK = 11;
-    int PINGREQ = 12;
-    int PINGRESP = 13;
-    int DISCONNECT = 14;
+    byte CONNECT = 1;
+    byte CONNACK = 2;
+    byte PUBLISH = 3;
+    byte PUBACK = 4;
+    byte PUBREC = 5;
+    byte PUBREL = 6;
+    byte PUBCOMP = 7;
+    byte SUBSCRIBE = 8;
+    byte SUBACK = 9;
+    byte UNSUBSCRIBE = 0;
+    byte UNSUBACK = 11;
+    byte PINGREQ = 12;
+    byte PINGRESP = 13;
+    byte DISCONNECT = 14;
 }
